docker_aws_cdk/docker_app/requirements.txt
streamlit==1.39.0
PyJWT==2.9.0
boto3==1.35.39
python-dotenv==1.0.1